+++ _test/test/goldens/generated_build_script.dart
@@ -154,12 +154,12 @@ final _builders = <_i1.BuilderApplication>[
     appliesBuilders: const [r'build_resolvers:transitive_digest_cleanup'],
   ),
   _i1.applyPostProcess(
-    r'build_modules:module_cleanup',
-    _i6.moduleCleanup,
-  ),
-  _i1.applyPostProcess(
     r'build_resolvers:transitive_digest_cleanup',
     _i8.transitiveDigestCleanup,
+  ),
+  _i1.applyPostProcess(
+    r'build_modules:module_cleanup',
+    _i6.moduleCleanup,
   ),
   _i1.applyPostProcess(
     r'build_web_compilers:dart2js_archive_extractor',