+++ build_test/lib/src/in_memory_reader_writer.dart
@@ -9,6 +9,8 @@ import 'package:build/build.dart';
 import 'package:build/src/internal.dart';
 // ignore: implementation_imports
 import 'package:build_runner_core/src/asset/reader_writer.dart';
+// ignore: implementation_imports
+import 'package:build_runner_core/src/generate/input_tracker.dart';
 import 'package:glob/glob.dart';
 
 import 'test_reader_writer.dart';
@@ -19,69 +21,66 @@ import 'test_reader_writer.dart';
 /// this package.
 class InMemoryAssetReaderWriter extends ReaderWriter
     implements TestReaderWriter {
+  /// Assets read directly from this reader.
+  final Set<AssetId> assetsRead;
+
   /// Create a new asset reader/writer.
   ///
   /// If provided [rootPackage] is the default package when globbing for files.
-  ///
-  /// Unlike the non-test [ReaderWriter], starts with an [inputTracker] so
-  /// inputs can be inspected after the test runs.
   factory InMemoryAssetReaderWriter({String? rootPackage}) {
     final filesystem = InMemoryFilesystem();
     return InMemoryAssetReaderWriter.using(
+      assetsRead: {},
       rootPackage: rootPackage ?? 'unset',
       assetFinder: InMemoryAssetFinder(filesystem, rootPackage),
       assetPathProvider: const InMemoryAssetPathProvider(),
       filesystem: filesystem,
       cache: const PassthroughFilesystemCache(),
-      inputTracker: InputTracker(),
     );
   }
 
   InMemoryAssetReaderWriter.using({
+    required this.assetsRead,
     required super.rootPackage,
     required super.assetFinder,
     required super.assetPathProvider,
     required super.filesystem,
     required super.cache,
-    required super.inputTracker,
-  }) : super.using();
+  }) : super.using() {
+    InputTracker.captureInputTrackersForTesting = true;
+  }
 
   @override
   InMemoryAssetReaderWriter copyWith({
     AssetPathProvider? assetPathProvider,
     FilesystemCache? cache,
   }) => InMemoryAssetReaderWriter.using(
+    assetsRead: assetsRead,
     rootPackage: rootPackage,
     assetFinder: assetFinder,
     assetPathProvider: assetPathProvider ?? this.assetPathProvider,
     filesystem: filesystem,
     cache: cache ?? this.cache,
-    inputTracker: inputTracker,
   );
 
   @override
   ReaderWriterTesting get testing => _ReaderWriterTestingImpl(this);
 
-  // Record all reads in `inputTracker` so tests can verify them.
-  //
-  // TODO(davidmorgan): refactor to remove differences in how test and real
-  // code use inputTracker; keep tracking for tests separate from real tracking.
-
   @override
   Future<bool> canRead(AssetId id) {
-    inputTracker!.assetsRead.add(id);
+    assetsRead.add(id);
     return super.canRead(id);
   }
 
   @override
   Future<List<int>> readAsBytes(AssetId id) async {
-    inputTracker!.assetsRead.add(id);
+    assetsRead.add(id);
     return super.readAsBytes(id);
   }
 
   @override
-  Future<String> readAsString(AssetId id, {Encoding encoding = utf8}) async {
-    inputTracker!.assetsRead.add(id);
+  Future<String> readAsString(AssetId id, {Encoding encoding = utf8}) {
+    assetsRead.add(id);
     return super.readAsString(id, encoding: encoding);
   }
 }
@@ -128,7 +127,13 @@ class _ReaderWriterTestingImpl implements ReaderWriterTesting {
       );
 
   @override
-  Iterable<AssetId> get assetsRead => _readerWriter.inputTracker!.assetsRead;
+  Iterable<AssetId> get inputsTracked =>
+      InputTracker.inputTrackersForTesting[_readerWriter.filesystem]!
+          .expand((tracker) => tracker.inputs)
+          .toSet();
+
+  @override
+  Iterable<AssetId> get assetsRead => _readerWriter.assetsRead;
 
   @override
   bool exists(AssetId id) => _readerWriter.filesystem.existsSync(