+++ build_resolvers/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Start using `package:build/src/internal.dart`.
 - Switch `BuildAssetUriResolver` dependency crawl to an iterative
   algorithm, preventing stack overflows.
+- Move `BuildStepImpl` to `build_runner_core`, use `SingleStepReader` directly.
 
 ## 2.4.4